Complete serpAPI coding challenge
To complete this challenge, I created a base parent class in ./site_format/site_structure.rb that has all the methods necessary to perform the scrapping of pages. The child classes Google, Bing and SerpApiChallenge that inherit from SiteStructure need to define instance variables according to how the websites they scrape are structured. The script to test my solution is in main.rb and can be run from the terminal. I also created specs for each scrapper that populates templates with data generated from the Faker gem to run the specs with dynamic data.